Fixed some warnings (part 7)

+++ lib86cpu/lib86cpu_priv.h
@@ -250,7 +250,7 @@ struct cpu_t {
 	addr_t db_addr;
 	addr_t instr_eip;
 	addr_t virt_pc;
-	size_t instr_bytes;
+	uint32_t instr_bytes;
 	uint8_t size_mode;
 	uint8_t addr_mode;
 	uint8_t translate_next;
